refactor: simplify pointer encoding API
The changes I've made here are subjective, but follow
conventional wisdom in some useful ways. I'd expect
Rust devs to find this more intuitive :)

First and foremost, we're documenting our invariants.
This is important for readability, as readers otherwise need
to internalise the entire API of our unsafe code to review it.
This way, we can reason locally at function boundaries.

The next most noticable change is in the `extract` function.
Because we wanted to use it on shared references, we
couldn't return a `Box` - `Box::from_raw` requires a mutable
pointer. There are a few ways to achieve this, and a
closure allows us to reuse the `match` without writing two
copies of `extract`.

I also replaced `forget` with `ManuallyDrop` in `slot`. This
is mostly stylistic: `forget` is soft-deprecated due to problems
with panics. The previous code was sound anyway.

This makes it easy to reuse the matching
logic. We  - the issue before was that `Box::from_raw` is not sound
on